exit py dev test

+++ py-dev-test.sh
@@ -1,15 +1,25 @@
 #!/usr/bin/env bash
 
-set -e
-
-function install_deps() {
-    python3 -m pip install -U pytest
+function install_deps
+{
+    
+    if ! python3 -m pip install -U pytest
+    then
+        (>&2 echo "🔥 We have failed to install `pytest`.")
+        exit 1
+    fi
 }
 
-function devtest {
-    python3 setup.py test
+function devtest
+{
+    if ! python3 setup.py test
+    then
+        (>&2 echo "🔥 Development test has failed.")
+        exit 1
+    fi
     git clean -xdfq
 }
 
 install_deps
 devtest
+echo "😊 Development test was a success."
